chore: Remove duplicate version from "all" (#38)
This action determines which versions of go we are supporting, including
a latest, penultimate, and a minimum version. When the minimum and
penultimate are the same version, we would have duplicate entries.

This would lead to multiple runs of the same sets of tests (which would
be fine), but failed when the test results were uploaded, due to naming
collisions.